+++ carp_mobile_sensing/test/cams_domain_test.dart
@@ -48,6 +48,17 @@ void main() {
       masterPhone, // a task with all measures
     );
 
+    // collect device info only once
+    masterProtocol.addTriggeredTask(
+        OneTimeTrigger('device'),
+        AutomaticTask()
+          ..measures = SamplingPackageRegistry().debug().getMeasureList(
+            types: [
+              DeviceSamplingPackage.DEVICE,
+            ],
+          ),
+        masterPhone);
+
     // adding two measures to another device
     masterProtocol.addTriggeredTask(
         ImmediateTrigger(),
@@ -67,10 +78,10 @@ void main() {
     expect(masterProtocol.ownerId, 'user@dtu.dk');
     expect(masterProtocol.masterDevices.length, 1);
     expect(masterProtocol.connectedDevices.length, 1);
-    expect(masterProtocol.triggers.length, 3);
+    expect(masterProtocol.triggers.length, 4);
     expect(masterProtocol.triggers.keys.first, '0');
-    expect(masterProtocol.tasks.length, 3);
-    expect(masterProtocol.triggeredTasks.length, 3);
+    expect(masterProtocol.tasks.length, 4);
+    expect(masterProtocol.triggeredTasks.length, 4);
   });
 
   test(